(feat): Authenticate comment routes

+++ api/routes/comment_routes.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 from flask import Blueprint, request, jsonify
-
+from flask_jwt_extended import get_jwt_identity, jwt_required
+from api.auth import role_required
 from api.models.picture import Picture
 from api.models.user import User
 from api.models.comment import Comment
@@ -13,6 +14,7 @@ comment_bp = Blueprint('main', __name__)
 
 # POST: Add a comment to a specific picture
 @comment_bp.route('/pictures/<int:picture_id>/comments', methods=['POST'])
+@jwt_required()
 def add_comment(picture_id):
     """add a comment to a picture"""
     data = request.get_json()
@@ -41,13 +43,15 @@ def add_comment(picture_id):
             "user_id": new_comment.user_id,
             "picture_id": new_comment.picture_id,
             "content": new_comment.content,
-            "timestamp": new_comment.timestamp
+            "created_at": new_comment.created_at,
+            "updated_at": new_comment.updated_at
         }
     }), 201
 
 
 # GET: Retrieve all comments for a specific picture
 @comment_bp.route('/pictures/<int:picture_id>/comments', methods=['GET'])
+@jwt_required()
 def get_comments(picture_id):
     """retrieves all comments for a specific picture"""
     picture = Picture.query.get(picture_id)
@@ -60,7 +64,8 @@ def get_comments(picture_id):
         "user_id": comment.user_id,
         "picture_id": comment.picture_id,
         "content": comment.content,
-        "timestamp": comment.timestamp
+        "created_at": comment.created_at,
+        "updated_at": comment.updated_at
     } for comment in comments]
 
     return jsonify(comments_list), 200
@@ -68,6 +73,7 @@ def get_comments(picture_id):
 
 # GET: Retrieve a comment by it's id
 @comment_bp.route('/comments/<int:comment_id>', methods=['GET'])
+@jwt_required()
 def get_comment(comment_id):
     """retrieve a comment by it's id"""
     comment = Comment.query.get(comment_id)
@@ -79,17 +85,24 @@ def get_comment(comment_id):
         "user_id": comment.user_id,
         "picture_id": comment.picture_id,
         "content": comment.content,
-        "timestamp": comment.timestamp
+        "created_at": comment.created_at,
+        "updated_at": comment.updated_at
     }), 200
 
 
 # PUT: Update a specific comment
 @comment_bp.route('/comments/<int:comment_id>', methods=['PUT'])
+@jwt_required()
 def update_comment(comment_id):
     """update a specific comment"""
     comment = Comment.query.get(comment_id)
     if not comment:
         return jsonify({"error": "Comment not found"})
+
+    # Only proceed to update if the current user owns the comment
+    current_user_id = get_jwt_identity()
+    if current_user_id != comment.user_id:
+        return jsonify({"error": "You are not allowed to update this comment"}), 403
 
     data = request.get_json()
     new_content = data.get('content')
@@ -99,6 +112,7 @@ def update_comment(comment_id):
 
     comment.content = new_content
     db.session.commit()
+    db.session.refresh(comment)
 
     return jsonify({
         "message": "Comment updated successfully!",
@@ -107,18 +121,25 @@ def update_comment(comment_id):
             "user_id": comment.user_id,
             "picture_id": comment.picture_id,
             "content": comment.content,
-            "timestamp": comment.timestamp
+            "created_at": comment.created_at,
+            "updated_at": comment.updated_at
         }
     }), 200
 
 
 # DELETE: Delete a specific comment
 @comment_bp.route('/comments/<int:comment_id>', methods=['DELETE'])
+@jwt_required()
 def delete_comment(comment_id):
     "delete a specific comment"
     comment = Comment.query.get(comment_id)
     if not comment:
         return jsonify({"error": "Comment not found"}), 404
+
+    # Only proceed to delete if the current user owns the comment
+    current_user_id = get_jwt_identity()
+    if current_user_id != comment.user_id:
+        return jsonify({"error": "You are not allowed to delete this comment"}), 403
 
     db.session.delete(comment)
     db.session.commit()
@@ -126,10 +147,11 @@ def delete_comment(comment_id):
     return jsonify({"message": "Comment deleted successfully!"}), 200
 
 
-# GET: Retrieve a user's comments by their id
-@comment_bp.route('/users/<int:user_id>/comments', methods=['GET'])
+# GET: Retrieve all the comments of a specific user
+@comment_bp.route('/user/<int:user_id>/comments', methods=['GET'])
+@role_required('admin')
 def get_user_comments(user_id):
-    """retrieve a specific user's comments"""
+    """retrieve all of a specific user's comments"""
     user = User.query.get(user_id)
     if not user:
         return jsonify({"error": "User not found"}), 404
@@ -140,7 +162,8 @@ def get_user_comments(user_id):
         "user_id": comment.user_id,
         "picture_id": comment.picture_id,
         "content": comment.content,
-        "timestamp": comment.timestamp
+        "created_at": comment.created_at,
+        "updated_at": comment.updated_at
     } for comment in comments]
 
     return jsonify(comments_list), 200
